+++ .storybook/preview.js
@@ -1,8 +1,8 @@
-import '@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js';
 import { setCompodocJson } from '@storybook/addon-docs/angular';
+import '@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js';
 import docJson from '../dist/documentation.json';
 // Needed for Orchestrator
-import 'core-js/features/reflect';
+import 'reflect-metadata/lite';
 
 setCompodocJson(docJson);