+++ libs/locale/data/rollup.config.mjs
@@ -1,9 +1,9 @@
-import typescript from '@rollup/plugin-typescript';
-import nodeResolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import { fileURLToPath } from 'url';
+import nodeResolve from '@rollup/plugin-node-resolve';
+import typescript from '@rollup/plugin-typescript';
 import { globbySync } from 'globby';
 import path from 'path';
+import { fileURLToPath } from 'url';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -50,17 +50,17 @@ export default locales
             dir: path.resolve(__dirname, locale, 'src/data'),
             format: 'es',
             chunkFileNames: '[name].ts',
+            sourcemap: false,
+            manualChunks: () => 'data',
         },
-        manualChunks: () => 'data',
         plugins: [
             nodeResolve(),
             commonjs(),
             typescript({
                 types: ['node', 'jest'],
-                target: 'es2015',
-                module: 'esnext',
                 exclude: ['**/*.spec.ts', '**/*.stories.ts'],
                 tsconfig: './tsconfig.base.json',
+                sourceMap: false,
             }),
         ],
     }));