+++ libs/locale/data/tr/src/index.ts
@@ -1,0 +1,13 @@
+import { I18nLocaleDataPackageModuleInlined } from '@spryker/locale';
+
+import * as inlineData from './data/data';
+
+export const data = Object.values<I18nLocaleDataPackageModuleInlined>(inlineData as any)
+    .filter((d) => d.data)
+    .reduce(
+        (acc, d) => ({
+            ...acc,
+            ...Object.fromEntries(Object.entries(d.data).map(([key, value]) => [`${d.name}.${key}`, value])),
+        }),
+        { locale: 'tr' },
+    );