+++ libs/locale/src/lib/locale-switcher/locale-switcher.stories.ts
@@ -2,9 +2,10 @@ import { importProvidersFrom } from '@angular/core';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
 import DeLocaleModule, { DE_LOCALE } from '@spryker/locale/locales/de';
 import EnLocaleModule, { EN_LOCALE } from '@spryker/locale/locales/en';
+import TrLocaleModule, { TR_LOCALE } from '@spryker/locale/locales/tr';
 import { LocaleModule } from '../locale.module';
 
-const locales = { English: EN_LOCALE, German: DE_LOCALE };
+const locales = { English: EN_LOCALE, German: DE_LOCALE, Turkish: TR_LOCALE };
 
 export default {
     title: 'LocaleSwitcherComponent',
@@ -14,6 +15,7 @@ export default {
                 importProvidersFrom(LocaleModule.forRoot()),
                 importProvidersFrom(EnLocaleModule),
                 importProvidersFrom(DeLocaleModule),
+                importProvidersFrom(TrLocaleModule),
             ],
         }),
         moduleMetadata({