+++ libs/table.filter.select/src/lib/types.ts
@@ -1,4 +1,4 @@
-// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+// eslint-disable-next-line @nx/enforce-module-boundaries
 import { TableFilterBase } from '@spryker/table.feature.filters';
 import { SelectValueSelected } from '@spryker/select';