+++ libs/tree-select/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-next.0](https://github.com/spryker/ui-components/compare/@spryker/tree-select@1.1.0...@spryker/tree-select@1.2.0-next.0) (2024-03-14)
+
+
+### Features
+
+* add turkish locale ([3dae079](https://github.com/spryker/ui-components/commit/3dae079b7b5be8affaeb2440d62ca0f750db715f))
+
+
+
+
+
 # [1.1.0](https://github.com/spryker/ui-components/compare/@spryker/tree-select@1.1.0-next.0...@spryker/tree-select@1.1.0) (2023-12-22)
 
 **Note:** Version bump only for package @spryker/tree-select